Particle - move particle elements to their own package
